all: Remove etcd
It has been replaced by an in-process Raft implementation in discoverd.
