--- Godeps/_workspace/src/github.com/coreos/go-etcd/etcd/version.go
@@ -1,3 +1,0 @@
-package etcd
-
-const version = "v2"
