--- appliance/etcd/Dockerfile
@@ -1,8 +1,0 @@
-FROM flynn/busybox
-
-ADD https://raw.githubusercontent.com/bagder/ca-bundle/master/ca-bundle.crt /etc/ssl/certs/ca-certificates.crt
-ADD ./bin/etcd /bin/etcd
-ADD ./bin/etcdctl /bin/etcdctl
-ADD ./bin/etcd-wrapper /bin/etcd-wrapper
-
-ENTRYPOINT ["/bin/etcd-wrapper"]
