--- appliance/etcd/build.sh
@@ -1,16 +1,0 @@
-#!/bin/bash
-
-set -eo pipefail
-
-version=2.0.11
-tmpdir=$(mktemp --directory)
-pkg="etcd-v${version}-linux-amd64"
-
-cleanup() {
-  rm -rf "${tmpdir}"
-}
-trap cleanup EXIT
-
-curl -L "https://github.com/coreos/etcd/releases/download/v${version}/${pkg}.tar.gz" | tar xzC "${tmpdir}"
-mkdir -p bin
-mv ${tmpdir}/${pkg}/{etcd,etcdctl} bin/
