+++ bootstrap/manifest_template.json
@@ -5,36 +5,6 @@
     "vars": ["CLUSTER_DOMAIN"]
   },
   {
-    "id": "etcd",
-    "app": {
-      "name": "etcd"
-    },
-    "action": "run-app",
-    "release": {
-      "env": {
-        "ETCD_INITIAL_CLUSTER_STATE": "new",
-        "ETCD_INITIAL_CLUSTER": "{{ range $ip := .SortedHostIPs }}{{ md5sum $ip }}=http://{{ $ip }}:2380,{{ end }}",
-        "DISCOVERD": "none"
-      },
-      "processes": {
-        "app": {
-          "ports": [{"port": 2379, "proto": "tcp"}, {"port": 2380, "proto": "tcp"}],
-          "data": true,
-          "resurrect": true,
-          "host_network": true,
-          "omni": true
-        }
-      }
-    },
-    "artifact": {
-      "type": "docker",
-      "uri": "$image_repository?name=flynn/etcd&id=$image_id[etcd]"
-    },
-    "processes": {
-      "app": 1
-    }
-  },
-  {
     "id": "flannel",
     "app": {
       "name": "flannel"
@@ -235,16 +205,6 @@
       "name": "postgres",
       "meta": {"flynn-system-app": "true"},
       "strategy": "postgres"
-    }
-  },
-  {
-    "id": "etcd-app",
-    "action": "add-app",
-    "from_step": "etcd",
-    "app": {
-      "name": "etcd",
-      "meta": {"flynn-system-app": "true"},
-      "strategy": "one-by-one"
     }
   },
   {
