--- discoverd/Procfile
@@ -1,2 +1,0 @@
-etcd: etcd -name tmp -data-dir /tmp -v
-discoverd: ./discoverd
