+++ router/Procfile
@@ -1,5 +1,4 @@
 # This Procfile boots router's development dependencies
-etcd: etcd
 discoverd: sleep 1; discoverd
 sample_backend: sleep 2; ./examples/server/server :9999
