+++ test/scripts/test-unit.sh
@@ -7,8 +7,8 @@ util/commit-validator/validate-gofmt
 
 bats script/test
 
-GIT_COMMIT=dev GIT_BRANCH=dev GIT_TAG=none GIT_DIRTY=false tup appliance/etcd discoverd
-PATH=$PATH:$PWD/appliance/etcd/bin:$PWD/discoverd/bin
+GIT_COMMIT=dev GIT_BRANCH=dev GIT_TAG=none GIT_DIRTY=false tup discoverd
+PATH=$PATH:$PWD/discoverd/bin
 
 export PGHOST=/var/run/postgresql
 sudo service postgresql start
